(chore): move es6 basic auth db seed logic to project root
- move db seed logic from src to project root
