+++ examples/es6/basic/seeder/mockData.js
